- making zView plugins compilable with gcc4 (yet untested, though) - removed PureC bits from zcodeclib and plugins (we can't mix PureC and gcc binaries even when same CDECL and int16 is used!) - unified env variables for CC, AR etc. for CROSS and native compilation - removed zlib (used system libz instead)

+++ zview/plugins/degas/degasldg.c
@@ -185,7 +185,7 @@ static int16 read_line( int16 bpl, uint8 *src, uint8 *compress_buffer)
     /* decompress the line */
     for( x = 0; x < bpl;)
     {
-        ( uint8)n = src[count++];
+        n = src[count++];
 
         if( n <= 127)           /* (n+1) literal bytes */
         {
